Change "has protocol" to "can call direct" and fix all paths.
IR protocol is not the only requirement for a direct call, and
the updating of both "has ir protocol" flag and the directly-
callable path was racy in both interpreter and JIT. In addition,
full builds and JIT that happen at threshold=0 in sync with
execution were sometimes continuing down the non-direct path using
the direct logic. This resulted in some call protocols running
twice, potentially messing up state for the direct logic.

The new flag is "canCallDirect", indicating that a transition to
the next tier has happened. Paths through normal non-direct logic
that might trigger a tier upgrade now re-check to see if that has
happened, re-branching back to the direct path.