When constructing branches during IRBuilder add some simple direct equality checks to eliminate the branch instr.  If we know we are comparing true to true then we also know we can just perform a jump.  This has an added big potential benefit that it should eliminate the next BB from the CFG entirely since it will no longer be reachable.
If two directly comparable values are never equal then the branch is replaced
with a noop.  This will probably not help at all once compiled down to bytecode
since the JVM will figure out we have a bogus conditional but it makes me
feel better (and it was driving me crazy to see 'while(true)' generate
'b_false(true)').