Multiple inlines in a single scope will work in a JIT'd method if they happen to occur in the same period.  IRScope has a boolean to disallow more than one inline to happen in a single scope but that does not mean multiple hot callsites might not happen to think they need to be inlined at the same time.
Do not assign a new callsiteID to a call so that profile calldata can actually
find the callsite.  Normally every time we do something new to a full build we
will clone it to make sure we do not step on anything potentially executing
code on a full build.  The callbase class will endless hand out new callsiteids.
So BB.clone will preserve the callsiteid.  This should probably change but
callsiteid most likely will replace ipc for finding callsite split so we will
increasingly depend on callsiteid.