New Site interface for getting (and for now setting) siteId.  All calls and yield currently use this id but this could be expanded for any types of sites we want to keep track of.
Using new Site we stop using ipc for find a call/yield instr which should
mean we do not need to generateInstructions in order to perform an inline
any more.

The main problem remaining with siteId is it is hard-wired into the
constructor to get a new one every time an object is created.  So we have
a setSiteId() method to restore the siteId on a cloned instance.  To fix this
we should make a clone constructor for all callbase and yield instrs.  This
is a bit painful so I am only mentioning it should be done at some point.