Speculative fix to crash for inlining:   def self.new(*)     super   end
This introduces a problem with ScopeModule not really having a scope to
figure out what its module is anymore (the scope was inlined away).  The
answer is to use the inlining candidate type and for non-zero ScopeModules it
is to subtract one from scopeDepth? (I am less sure about this).  With this
change, self.new still does not properly inline but this seems like a step
in the right direction (and I want @subbuss to see the approach).