Remove buffering for input source.
If it's an inline script, it is in the form of a byte[] and does
not need to be buffered.

If it's a file or stdin, it will be read via a RubyIO, which does
its own buffering.

If it's coming from a jar file, it's already buffered or will be
buffered by RubyIO.

Removing this buffering enables direct access to any FileChannel
associated with the open file, which in turn provides a true DATA
object when requested.