Cleaner attempt at #3577.
The problems in #3577 stemmed from the introduction of a second
exit instruction into global ensure blocks (rethrow saved exc).
Logic elsewhere in our passes that dealt with the GEB all assumed
that there would be exactly one control-transfer instruction at
the end of the GEB, and inserted various logic before this point:

* Stores of modified variables into the current binding.
* Pops of binding and frame at the end of the method body.

Once there were two instructions, it becamse a mess to know where
to insert these instructions.

This change modifies the "rethrow" instruction to itself be a
special case of "return" that may throw a saved exception. In both
cases, it is unrolling at least the current stack frame, and this
allows us to have a single instruction to do both normal returns
and saved exception rethrows in the GEB.

This change reverts previous fixes for #3577 that attempted to
work around the double-exit.