Add a new pretty-printing IR dumper.
The goal here is to keep the dumper as simple as possible, so I've
started out using reflection APIs to walk internal state in the
IR rather than relying on one-off methods in each Instr.

The reflective field walk will skip any fields marked as
transient. This is perhaps an unusual overloading of the Java
transient keyword, but I justify it thusly:

* It's cleaner than introducing and querying an annotation.
* It is at least somewhat related to externalization of an
  object graph.
* We have no intention to EVER use automatic serialization for
  our IR.
* Maintaining a separate map of which fields to display would
  have been error-prone and required custom code in each Instr.

The format is designed to be as readable as possible, so I set it
up to center on assignment to temp vars (registers) since that is
what throws off my reading of current IR dumping more than any
other factor.

ANSI color is supported to help different types of IR object pop
a bit more.