Add property and non-volatile table-based ivars.
This is in prep for officially saying we don't guarantee that
ivarrs will be volatile. Users concerned about this are
recommended to look into concurrent-ruby, which provides better
guarantees and utilities for volatility, atomicity, and more.

Note that there may still be a chance that other volatile accesses
around the variable updates force it to be effectively volatile,
but this will not be the case when reifying variables to be Java
fields, as we hope to do in an upcoming release.

See discussions surrounding https://github.com/ruby-concurrency/concurrent-ruby/pull/422
for more details.

See also #3353 for discussion about get volatility prior to this
upcoming change in policy.