[Truffle] Added argument validation for String#num_bytes=.
Now that we use ropees, our usages of this method should drop off. We currently only support truncating a String via this method. We no longer support expanding the underlying byte[] since ropes are immutable, so there's nothing to be gained by allocating extra storage. Any places relying on that mechanism to reserve space to write to should be updated to use a rope-friendly technique.