Parser changes for #2960: Block Parameter Scoping (vs Ruby).
A followup commit will try and address the depth change hacks in IR
but we still need to create some variables for magic vars in evals
and for flip flop.  So unfornately just deleting that fix code in
IR breaks those things.  Leaving the IR code does not break our
compat so it more of a maintenance issue...