Some more frozen-literal support.  This hooks command line options to the parser. There is still some problems to work through which I think is partially from the last commit being incomplete (about our missing lit node type) but almost all of the MRI tests are passing now.
Next up will be updating ripper to be 2.3 grammar and probably a little
refactoring to make this frozen literal stuff fit in better.

The one missing feature within this feature is --debug-frozen-string-literal.
This extra feature records where the frozen string literal was created.  So
to implement this we will need to somehow record the position into the string
so we can report it's original location in the freeze error message.  I am not
sure how important that will be for 2.3 support but it is still outstanding.