Preliminary support for debug-frozen-string-literal.
This was done in easiest way possible in that IR just does a put_field
onto the string literal with its location into @debug_created_info.
MRI also stores into an ivar but it is a numeric slot so it is not
user visible.  Without actually modifying IR this seemed the closest
mechanism.  It also has an interesting benefit that you can write
Ruby to access it and make your own tooling.  I put a FIXME: about
whether we should make this user-visible or not.  If we decide this
is a nice user-visible feature we should make a RFE on redmine.

The other slightly oddity of this impl is a dyndispatch to "freeze"
in IR since we have no instr for freezing.  I think though we could
possibly augment put_field to have a boolean which allows us to
by-pass freeze and then we could make a frozen literal and still
put field on it.

One other mechanism I thought about would be if we could make an
instr for dataGetStruct() for true non-ruby-visible storage.  This
seems like it would lead to a lot of new instrs but if we want to
manipulate native data maybe we need something in IR for that?

All-in-all this is a debugging feature and I do not think we should
worry too much about implementation so long as it gives the expected
output...

New options: --debug={list} --debug-frozen-string-literal.