Remove unconditional warning here.
I believe this was left in from dev time when we were still
sorting out how to treate evals with blocks. The common case that
produces the warning is from #3541, when a Symbol#to_proc is used
for an instance_exec or instance_eval. These are valid uses of
the proc and work correctly (using the receiver as the "self" for
the proc's symbol dispatch).

Because all the remaining block bodies that would trigger this
warning are Java-based blocks without any concept of frame/scope
binding, and since they still honor the incoming self, I'm going
to just remove the warning.

Fixes #3541.