Improve constant caching and factor logic out.
ConstantLookupSite now handles all constant lookup cases:

* Lexical+hierarchical search in searchConst
* Hierarchical search in inheritanceSearchConst
* LexicalSearch in lexicalSearchConst (for defined?)

Up to now, all three forms would cache one recent constant and
replace that binding on invalidation. This is obviously slow
when there are more than one possible constant value (lookups in
different hierarchies, for example) so I have modified it to chain
a series of bindings. This chain is currently not limited, and
that should be fixed; but the current logic improves red/black by
around 12%.