Use pure-Ruby Enumerable#inject with new block arg instr.
This is a modified version of the same inject used in Truffle,
but with a different form for the "single block arg" instruction.

With the previous (broken) version of this code using |o,|, I saw
a significant improvement in the performance of rubykon's bench
"full_playout.rb". Unfortunately with this new fixed version, that
improvement has disappeared and I have not figured out why. It is
possible that the addition of the fixed path, which allocates an
Array for more than two args passed, has now taken all the gains
back again.