[Truffle] Remove symbol helpers from RubyContext.
I'm trying to make RubyContext less of a God-object, and make it clearer
who has responsibility for what. So I'd rather than people get the
SymbolTable, and then get a Symbol from that, rather than giving the
impression that RubyContext is also responsible for this and routing
so much through that one class.