Make RubySymbol EncodingCabable so unmarshaling can set encoding.
This makes me a little nervous, in that we're modifying a symbol
looked up from the global table, and that could cause it to
overwrite an existing symbol's encoding. However, the way symbol
encodings are marshaled does not allow us to easily get the
encoding right away. Will have to experiment and see if there's
a real issue here.