Move debug logic for frozen strings out of IR.
The original logic built and installed the debug info as a series
of IR instructions. This allowed adding the feature without having
to modify JIT, but was somewhat heavy in code. This change moves
the debug check into the logic for FrozenString itself, and more
efficiently compiles it into bytecode.