[Truffle] Pass the correct encoding through to Rubinius::Mirror::String#splice.
Prior to this change, it was expected the caller would splice a string and then set the correct encoding for the final product. In the time in between, the String could be invalid. When we simply modified the underlying byte[] this was an acceptable trade-off. However, now that we're using ropes, it's possible the splice operation ends up using a bad specialized version of a rope operation, as the rope operations are encoding-aware.