Fix block jitting by forcing full build before checking protocol.
Fixes #3626.

I added code before 9.0.5.0 to ensure we only compiled blocks that
had full block protocol, since I don't want the JIT paths to have
by-hand protocol logic (interferes with call sites and indy opto.)
Unfortunately I checked the call protocol flag without forcing a
full build, so it was always false, so 9.0.5.0 was not ever
jitting blocks.

This fix forces a full JIT build of IR before checking for call
protocol on blocks.

This check may be unnecessary, if we *always* apply call protocol
to blocks now. I did not check whether that might be true, and
the only down side to this change is that more IR work happens
in the synchronous call thread, and if the build fails due to an
error it will be more fatal.