Fixes #3745. IO#block_nonwrite causes CCE on Windows.
I made a marker interface so that both ChannelStream and CRLFStream
wrapper can have a common type so they can call writenonblock.  I added
an error check to throw and IOException if CRLFStreamWrapper is wrapping
from stream type but I suspect it is an impossible path by this point in
our code.  Zero 1.7 risk fix since this was just dumping stack before.