Hello new instr SearchModuleForConst.  This instr is somewhat of a combo instr.  It used to be emitted as InheritanceSearch + BNE + ConstMissing instr but this ended up being undesirable for a few or reasons:
1. number of overall instrs are larger for no benefit.  IR as it
exists could not make use of the finer granularity and based on
how many side-effecting things get called between constant
lookups it was unlikely we will be able to simplify a scope to
only lookup a constant once.  Also under the IR layer can change
const resolution to use the same invokedynamic bindings for the
same "constant" in a scope essentially gambling that all constants
with the same name will end up referencing the same constant in
a given IR scope (which is probably 99.99999% true -- prove this
wrong! :) ).

2. increases the size of the CFG for no benefit.  not as big
of a deal...

3. interpretation is a little bit faster.  we moved this logic
into Java so we are no longer emulating if statement and doing
less overall work.  I tried red black bench and saw a small
but measurable increase.