[Truffle] Removed caching of ByteList in Rope.
Most ropes never need a ByteList, so paying the cost of the reference in every rope was not ideal. Moreover, most repeated usages of the Rope -> ByteList adapter (those that would benefit from a cached object) are found in the regexp subsystem and all of that is slow path currently. Thus, removing this reference should save RAM and have negligible impact on performance.