[Truffle] Added specializations for composite ropes in RopeNodes.GetByteNode.
Generally we can't optimize much on the structure of a composite rope, as they're
recursive data structures and we can't PE unbounded recursion. Instead, we must
find terminal points in the rope that we can optimize on, such as whether the raw
bytes reference is populated.

This change introduces deconstructing composite ropes for a single level. There
are various use cases in Ruby that result in a LeafRope being added as a descendant
to a composite rope, which immediately breaks many fast paths. Examples include:

* Appending "\n" to a LeafRope, resulting in a ConcatRope.
* Chopping "\n" from a LeafRope, resulting in a SubstringRope.
* Using a LeafRope as a pattern to fill out a string, resulting in a RepeatingRope.

While this won't make arbitrary rope constructions faster, it will handle a common
use case.