[Truffle] All rope bytes will start at offset 0.
The code that was here was placeholder code for algorithms converted from ByteList. The idea was something like SubstringRope could share its raw bytes with its child, and then alter its 'begin' value to demarcate its real portion of the underlying byte array. While there may be real benefits to that level of byte sharing, it complicates all code that must work with the bytes. And we've already assumed the 'begin' value is 0 in several places.

Should we wish to revisit the byte sharing, it's probably best done in the context of a specialized node for whatever operation.