Search all scopes in turn for refinements.
We were not properly handling the case of a method starting in a
scope that is not a cref (class/module/script) body. The old logic
would grab the current scope, find no refinements, and then move
to the previous cref scope. This unfortunately would jump out of
the current lexical cref into the containing lexical cref and miss
any refinements in the current class body.

We may be able to improve this by only searching cref scopes.

Fixes #3548. See that bug for a reproduction.