[Truffle] Fixed loading core files on Windows.
My previous fix for relative path names broke things on Windows. The problem is canonicalizing a path makes it fully qualified. On Linux & Mac, that's not a problem because the classpath root and the filesystem root use the same charact ('/'), but on Windows the filesystem root is 'C:\'. We need to normalize the path instead and then replace any backslashes with forward slashes to make it a valid path for the resource loader.