Re-port Array#hash and some related logic to fix #3884.
It appears MRI defers recursion detection to the elements of an
array and has other logic for combining hashes of those elements.
I also added better initial hash value similar to MRI's use of a
pointer in memory, which appears to be the main thing helping this
issue, since now the hash of an empty array is a useful number.