Fixes #3680.   JRuby and devise Null pointer exception
This was largely just some code which was added before prepend existed.
It would re-wrap begin and end of a range and make an array and call min
on that.  This would obviously not have anything prepended on it and it
would not even be the same type.

I ended up re-porting min and max in range from MRI since I found another
mistake with min(a) and max(a) not passing a to super if invoked in the
block form.  This also makes us a tiny bit faster since we end up only
doing on cmp in cases of fixnums (why is this even optimized by MRI :) ).