[Truffle] Remove Rubinius README
All files that have come from Rubinius or the rubinius-core-api project
are individually licensed. We don't need this README file to tell us
where they have come from or what their licenses are. Versions are
not relevant any more, as we're not pulling from source in the future.

As well as being redundant, this README won't really fit anywhere as we
start to absorb Rubinius code into a unified code base (maintaining
license headers).

The historical information of course is maintained in this commit.