Test out MRI's timeout.rb unmodified to fix #3834.
Our native Timeout has not been updated in a long time, and for

* A separate anonymous exception is not used to trigger the
  timeout anymore; instead, they use catch/throw.
* The exception message string is pre-allocated and frozen.
* Various other miscellaneous changes.

This version has *not* been tested for performance, and it will
initially be much slower than our scheduled executor-based version
due to it always spinning up a timeout thread. But we can see how
it goes in test suites and then evaluate how to move forward.

Note that the change to test/jruby/test_timeout.rb was to bring
behavior in line with MRI 2.3.