Fixes to get LocalVariable compiling in all contexts.
Many places just used jvmStoreLocal to store the variable, which
assumed (because we only ran with call protocol in place) that
all such stores would be to Java locals. I refactored this method
to support LocalVariable as well as a different form that avoids
stack-juggling to insert the value into the scope. This appears to
get almost all code compiling that compiled before.