Temporarily pass env values through bytes for locale encoding.
This fixes the one regression in my previous changes. Because MRI
always creates new strings when you read from ENV, they just
normalize to locale encoding for setenv, and then either use
locale encoding or filesystem encoding for getenv. The equivalent
for us here is to just let the setenv string decode to Java String
and then use the default getBytes to match it up with the default
Charset used for locale, rather than attempting to transcode the
string directly.