Proof-of-concept for a flatbuffer-based interpreter.
This is a first attempt at a buffer that uses flatbuffers to
execute directly from disk. Enough instructions and operands are
implemented for a simple "hello world" but there's a few corners
cut for simplicity.

Much will need to be improved for this to be viable, but it shows
promise that I could get something working in a basic form this
quickly.

* Most tables from the buffer are created new every time. A better
  approach would be to have a pool of these table objects that
  can be reused within a given interpreter frame without
  reallocating on every entry or every operation.
* I have glossed over some details of how calls work, but the
  basics are there.
* This implementation will lead to the "big switch" problem we
  have seen impact straight-line performance in previous
  interpreters. However, since this is not intended to be the
  "fast" way to run Ruby code (the goal is starting up without
  parsing and compiling Ruby code) it may not be important.

Here's an example run:

```
$ java -Djruby.home=`pwd` -cp lib/jruby.jar org.jruby.ir.persistence.flat.FlatIRInterpreter 'puts "Hello, flatbuffers!"'
Hello, flatbuffers!
```