Use new visitor logic everywhere.
Nearly all of these are now allocated once, like a lambda. A few
require more state to be passed in than is reasonable for a
generic visitor, or require state to be calculated during visit
and retrieved at the end. These cases were at least reduced to
only allocate a visitor, rather than a visitor and a state-holding
object.