Removed bad entries from jruby-complete.jar.
I think this was intended to put the license files in the META-INF/ path. Unfortunately, it seems to create a path that mirrors the host filesystem in the generated JAR, rather than placing the files in the META-INF/ path in the JAR. Moreover, the licenses are placed in META-INF/ by some other process, so even if it work, I don't think this config section is required.