Unify and clean up dispatch paths for Java methods and ctors.
This work was prompted by running test:jruby:jit with indy turned
on, which caused many calls to follow the IRubyObject[] path where
they would have followed the specific-arity paths. Failures showed
up because the IRubyObject[] paths did not handle error messages
the same way as the specific paths.

So I started by modifying findCallable to switch on arity, to use
the same logic for low-arity calls whether they come through
the appropriate path or not. That caused more bugs because there
were expectations in the tests that varargs processing would be
done for all of the IRubyObject[] path and the specific-arity
findCallable versions did not do varargs.

So I added varargs logic to all paths except for arity = 0,
thinking it did not need it. That broke a test that expected the
no-arg constructor to be called rather than the vararg constructor
when called with no arguments, so I added varargs logic to that
as well.

All tests appear to pass. I have some minor concern about the
additional overhead of checking for varargs matches, but not doing
it before was inconsistent, and we will only look for varargs
matches when nothing else matches and there's actually varargs
methods to look at.

Please review my changes, @kares.