Implement fully-cached "checked" methods and type conversions.
This batch of changes implements cached logic for the "checked"
type conversions and calls, and wires that logic up for many
places in core. CheckedSites is introduced into JavaSites as a
carrier for the four call site types needed to do a checked
invocation. With this change, checked calls can now be fully
cached.

This also introduces a few new cached unchecked paths and wires
them up in a few places as well. I will let this bake and then
ensure all places using the old logic now use the new logic in an
uncoming commit.