Roughed in IGV output for a CFG.  Next will be connecting it to an option and being able to also feed it via a socket connection.  It is missing graph tag.
Next steps in output will be for compiler passes to show which instrs were
removed and the same for CFG data.  Then we should be able to see how a pass
changes things.  We can also consider dumping LVA data for viewing too.

A supplemental enhancement in this commit is to add JumpTargetInstr
interface so we can ask a Branch of Jump for its jump target.