Remove RPC from instr.  Remove unused interpreter engines (at least for now).
The RPC not resides in a table in the FIC.  It is looked up via a list
of pairs.  I have not measured performance but O(n) compared vs O(1)
where n is the number of BBs (which is generally fairly small).  If
we see any slowdown this can be changed to O(ln(n)).