Add FIXME and link to issue 4051 to RETURN and END trace instructions
Unlike the LINE and CALL trace instructions, currently the parser does
not supply line numbers for RETURN and END trace instructions.

This is worked around by by the TraceInstr by computing line numbers
dynamically, rather than statically, but the same approach cannot be
applied to line number instructions, which means that line numbers in
stack traces generated inside the trace function will not be correct.

As the fix seems to be quite hard, I think it's important to document
this bug.

Issue #4051