Fixes #4051  Wrong line numbers in Kernel#caller inside set_trace_func (with code to reproduce).  More specifically this is a followup to @ivoanjo fixing the call and line events.  This round fixes up return events.
It is unfortunate, but there is a distinct lack of tests and no
specs for set_trace_func.  We do not normally run MRI tests on this
as we do not pass any c-* events and apparently do not do tracepoint.
I did run all the tests MRI has for set_trace_func and we only
failed for missing c-call and c-return events.  If I was not so
busy this month I would spend some time making a nicer test suite
for this.

There is a tiny tiny chance I will break some API due to changing
DefNode interface.  I found the liklihood of this so small I went
with it.  If I break someone you can internet shame me about it
later :|