Synchronize around getgrgid call. Fixes #4057.
getgrgid is thread-safe on OS X, but other platforms are not
required to be thread-safe by the POSIX specification. It appears
that at least some flavors of Linux do not have thread-safe
getgrgid. To aid users who might be calling it across threads, we
lock around the native getgrgid call.

I chose to lock around the RubyEtc.class to ensure the native call
is single-threaded for all JRuby instances in a given classloader.
There's no clean way to make that guarantee across classloaders.