Add classpath:/ as an absolute path prefix. See #4000
This does not appear to be necessary in 1.7, but in 9k if I do not
do this then require 'classpath:/somefile' fails to find the file.
Stepping through it, I see it skips the absolute path logic that
that would have handled it properly and proceeds on to load path
logic, which obviously does not find the file.