If parent class was reifiable, reify class. Fixes #3950.
The problem here was that if only the parent reified, the child's
map of instance variable will use the wrong accessors. With
invokedynamic this led to both parent and child putting ivars in
different places.

This change reifies child classes of other reifiable classes,
where before only direct descendants of Object would reify.