Can't reference Frame.DUMMY due to circular static init.
Frame has a static DUMMY = new Frame(). This causes Block's static
init to run, which has a DUMMY = new Block(). This causes
Binding's static init to run, which has a DUMMY = new Binding that
uses Frame.DUMMY. But we're still in the static init for Frame, so
Frame.DUMMY is still null.

Frame.DUMMY is just a utility for other areas, so creating a new
blank dummy frame for Binding.DUMMY is ok.